+++ native/main.cpp
@@ -19,6 +19,7 @@ public:
             s.enabled=number(0,1)!=0;s.freeCamera=number(0,1)!=0;s.targeting=number(0,1)!=0;
             s.crosshair=number(0,2);s.delayMs=number(0,1000);s.coneDegrees=number(0,90);
             s.aimAssist=number(0,1)!=0;s.assistStrength=number(0,80);s.debugLogging=number(0,1)!=0;
+            s.cameraMode=number(0,2);s.trackingSpeed=number(10,100);s.trackingResumeMs=number(0,3000);
             CombatCamera::configure(s);return 0;
         });
         lua.register_function("_CCStart",[](const Lua& l){